refactor proposal detail router
