+++ src/app/routers/proposals.py
@@ -1,9 +1,6 @@
-from fastapi import APIRouter, Path, Query
+from fastapi import APIRouter, Query
 
-from app.contracts.proposals import (
-    ProposalDetailEnvelopeResponse,
-    ProposalListEnvelopeResponse,
-)
+from app.contracts.proposals import ProposalListEnvelopeResponse
 from app.middleware.correlation import correlation_id_var
 from app.services.advisory_service_provider import proposal_service
 
@@ -70,30 +67,3 @@ async def list_proposals(
         "cursor": cursor,
     }
     return await service.list_proposals(filters=filters, correlation_id=correlation_id)
-
-
-@router.get(
-    "/{proposal_id}",
-    response_model=ProposalDetailEnvelopeResponse,
-    summary="Get Proposal",
-    description="Returns the latest proposal envelope for a specific advisory proposal id.",
-)
-async def get_proposal(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    include_evidence: bool = Query(
-        default=False,
-        description="Whether to request proposal evidence and support metadata when available.",
-        examples=[True],
-    ),
-) -> ProposalDetailEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_proposal(
-        proposal_id=proposal_id,
-        include_evidence=include_evidence,
-        correlation_id=correlation_id,
-    )
